# Optimize some Matrix4x4 operations with SSE