--- a/src/System.Numerics.Vectors/src/System/Numerics/Matrix4x4.cs
+++ b/src/System.Numerics.Vectors/src/System/Numerics/Matrix4x4.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Globalization;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Intrinsics.X86;
 
@@ -2086,42 +2087,7 @@ namespace System.Numerics
         /// </summary>
         /// <param name="value">The source matrix.</param>
         /// <returns>The negated matrix.</returns>
-        public static unsafe Matrix4x4 operator -(Matrix4x4 value)
-        {
-            if (Sse.IsSupported)
-            {
-                var zero = Sse.SetZeroVector128();
-                Sse.Store(&value.M11, Sse.Subtract(zero, Sse.LoadVector128(&value.M11)));
-                Sse.Store(&value.M21, Sse.Subtract(zero, Sse.LoadVector128(&value.M21)));
-                Sse.Store(&value.M31, Sse.Subtract(zero, Sse.LoadVector128(&value.M31)));
-                Sse.Store(&value.M41, Sse.Subtract(zero, Sse.LoadVector128(&value.M41)));
-
-                return value;
-            }
-            else
-            {
-                Matrix4x4 m;
-
-                m.M11 = -value.M11;
-                m.M12 = -value.M12;
-                m.M13 = -value.M13;
-                m.M14 = -value.M14;
-                m.M21 = -value.M21;
-                m.M22 = -value.M22;
-                m.M23 = -value.M23;
-                m.M24 = -value.M24;
-                m.M31 = -value.M31;
-                m.M32 = -value.M32;
-                m.M33 = -value.M33;
-                m.M34 = -value.M34;
-                m.M41 = -value.M41;
-                m.M42 = -value.M42;
-                m.M43 = -value.M43;
-                m.M44 = -value.M44;
-
-                return m;
-            }
-        }
+        public static Matrix4x4 operator -(Matrix4x4 value) => Negate(value);
 
         /// <summary>
         /// Adds two matrices together.
@@ -2129,40 +2095,7 @@ namespace System.Numerics
         /// <param name="value1">The first source matrix.</param>
         /// <param name="value2">The second source matrix.</param>
         /// <returns>The resulting matrix.</returns>
-        public static unsafe Matrix4x4 operator +(Matrix4x4 value1, Matrix4x4 value2)
-        {
-            if (Sse.IsSupported)
-            {
-                Sse.Store(&value1.M11, Sse.Add(Sse.LoadVector128(&value1.M11), Sse.LoadVector128(&value1.M11)));
-                Sse.Store(&value1.M21, Sse.Add(Sse.LoadVector128(&value1.M21), Sse.LoadVector128(&value1.M21)));
-                Sse.Store(&value1.M31, Sse.Add(Sse.LoadVector128(&value1.M31), Sse.LoadVector128(&value1.M31)));
-                Sse.Store(&value1.M41, Sse.Add(Sse.LoadVector128(&value1.M41), Sse.LoadVector128(&value1.M41)));
-                return value1;
-            }
-            else
-            {
-                Matrix4x4 m;
-
-                m.M11 = value1.M11 + value2.M11;
-                m.M12 = value1.M12 + value2.M12;
-                m.M13 = value1.M13 + value2.M13;
-                m.M14 = value1.M14 + value2.M14;
-                m.M21 = value1.M21 + value2.M21;
-                m.M22 = value1.M22 + value2.M22;
-                m.M23 = value1.M23 + value2.M23;
-                m.M24 = value1.M24 + value2.M24;
-                m.M31 = value1.M31 + value2.M31;
-                m.M32 = value1.M32 + value2.M32;
-                m.M33 = value1.M33 + value2.M33;
-                m.M34 = value1.M34 + value2.M34;
-                m.M41 = value1.M41 + value2.M41;
-                m.M42 = value1.M42 + value2.M42;
-                m.M43 = value1.M43 + value2.M43;
-                m.M44 = value1.M44 + value2.M44;
-
-                return m;
-            }
-        }
+        public static Matrix4x4 operator +(Matrix4x4 value1, Matrix4x4 value2) => Add(value1, value2);
 
         /// <summary>
         /// Subtracts the second matrix from the first.
@@ -2170,40 +2103,7 @@ namespace System.Numerics
         /// <param name="value1">The first source matrix.</param>
         /// <param name="value2">The second source matrix.</param>
         /// <returns>The result of the subtraction.</returns>
-        public static unsafe Matrix4x4 operator -(Matrix4x4 value1, Matrix4x4 value2)
-        {
-            if (Sse.IsSupported)
-            {
-                Sse.Store(&value1.M11, Sse.Subtract(Sse.LoadVector128(&value1.M11), Sse.LoadVector128(&value1.M11)));
-                Sse.Store(&value1.M21, Sse.Subtract(Sse.LoadVector128(&value1.M21), Sse.LoadVector128(&value1.M21)));
-                Sse.Store(&value1.M31, Sse.Subtract(Sse.LoadVector128(&value1.M31), Sse.LoadVector128(&value1.M31)));
-                Sse.Store(&value1.M41, Sse.Subtract(Sse.LoadVector128(&value1.M41), Sse.LoadVector128(&value1.M41)));
-                return value1;
-            }
-            else
-            {
-                Matrix4x4 m;
-
-                m.M11 = value1.M11 - value2.M11;
-                m.M12 = value1.M12 - value2.M12;
-                m.M13 = value1.M13 - value2.M13;
-                m.M14 = value1.M14 - value2.M14;
-                m.M21 = value1.M21 - value2.M21;
-                m.M22 = value1.M22 - value2.M22;
-                m.M23 = value1.M23 - value2.M23;
-                m.M24 = value1.M24 - value2.M24;
-                m.M31 = value1.M31 - value2.M31;
-                m.M32 = value1.M32 - value2.M32;
-                m.M33 = value1.M33 - value2.M33;
-                m.M34 = value1.M34 - value2.M34;
-                m.M41 = value1.M41 - value2.M41;
-                m.M42 = value1.M42 - value2.M42;
-                m.M43 = value1.M43 - value2.M43;
-                m.M44 = value1.M44 - value2.M44;
-
-                return m;
-            }
-        }
+        public static Matrix4x4 operator -(Matrix4x4 value1, Matrix4x4 value2) => Subtract(value1, value2);
 
         /// <summary>
         /// Multiplies a matrix by another matrix.
@@ -2211,69 +2111,7 @@ namespace System.Numerics
         /// <param name="value1">The first source matrix.</param>
         /// <param name="value2">The second source matrix.</param>
         /// <returns>The result of the multiplication.</returns>
-        public static unsafe Matrix4x4 operator *(Matrix4x4 value1, Matrix4x4 value2)
-        {
-            if (Sse.IsSupported)
-            {
-                Matrix4x4 m = default;
-
-                Sse.Store(&m.M11,
-                    Sse.Add(Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M11), Sse.LoadVector128(&value2.M11)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M12), Sse.LoadVector128(&value2.M21))),
-                            Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M13), Sse.LoadVector128(&value2.M31)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M14), Sse.LoadVector128(&value2.M41)))));
-
-                Sse.Store(&m.M21,
-                    Sse.Add(Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M21), Sse.LoadVector128(&value2.M11)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M22), Sse.LoadVector128(&value2.M21))),
-                            Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M23), Sse.LoadVector128(&value2.M31)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M24), Sse.LoadVector128(&value2.M41)))));
-
-                Sse.Store(&m.M31, 
-                    Sse.Add(Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M31), Sse.LoadVector128(&value2.M11)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M32), Sse.LoadVector128(&value2.M21))),
-                            Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M33), Sse.LoadVector128(&value2.M31)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M34), Sse.LoadVector128(&value2.M41)))));
-
-                Sse.Store(&m.M41,
-                    Sse.Add(Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M41), Sse.LoadVector128(&value2.M11)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M42), Sse.LoadVector128(&value2.M21))),
-                            Sse.Add(Sse.Multiply(Sse.SetAllVector128(value1.M43), Sse.LoadVector128(&value2.M31)),
-                                    Sse.Multiply(Sse.SetAllVector128(value1.M44), Sse.LoadVector128(&value2.M41)))));
-
-                return m;
-            }
-            else
-            {
-                Matrix4x4 m;
-
-                // First row
-                m.M11 = value1.M11 * value2.M11 + value1.M12 * value2.M21 + value1.M13 * value2.M31 + value1.M14 * value2.M41;
-                m.M12 = value1.M11 * value2.M12 + value1.M12 * value2.M22 + value1.M13 * value2.M32 + value1.M14 * value2.M42;
-                m.M13 = value1.M11 * value2.M13 + value1.M12 * value2.M23 + value1.M13 * value2.M33 + value1.M14 * value2.M43;
-                m.M14 = value1.M11 * value2.M14 + value1.M12 * value2.M24 + value1.M13 * value2.M34 + value1.M14 * value2.M44;
-
-                // Second row
-                m.M21 = value1.M21 * value2.M11 + value1.M22 * value2.M21 + value1.M23 * value2.M31 + value1.M24 * value2.M41;
-                m.M22 = value1.M21 * value2.M12 + value1.M22 * value2.M22 + value1.M23 * value2.M32 + value1.M24 * value2.M42;
-                m.M23 = value1.M21 * value2.M13 + value1.M22 * value2.M23 + value1.M23 * value2.M33 + value1.M24 * value2.M43;
-                m.M24 = value1.M21 * value2.M14 + value1.M22 * value2.M24 + value1.M23 * value2.M34 + value1.M24 * value2.M44;
-
-                // Third row
-                m.M31 = value1.M31 * value2.M11 + value1.M32 * value2.M21 + value1.M33 * value2.M31 + value1.M34 * value2.M41;
-                m.M32 = value1.M31 * value2.M12 + value1.M32 * value2.M22 + value1.M33 * value2.M32 + value1.M34 * value2.M42;
-                m.M33 = value1.M31 * value2.M13 + value1.M32 * value2.M23 + value1.M33 * value2.M33 + value1.M34 * value2.M43;
-                m.M34 = value1.M31 * value2.M14 + value1.M32 * value2.M24 + value1.M33 * value2.M34 + value1.M34 * value2.M44;
-
-                // Fourth row
-                m.M41 = value1.M41 * value2.M11 + value1.M42 * value2.M21 + value1.M43 * value2.M31 + value1.M44 * value2.M41;
-                m.M42 = value1.M41 * value2.M12 + value1.M42 * value2.M22 + value1.M43 * value2.M32 + value1.M44 * value2.M42;
-                m.M43 = value1.M41 * value2.M13 + value1.M42 * value2.M23 + value1.M43 * value2.M33 + value1.M44 * value2.M43;
-                m.M44 = value1.M41 * value2.M14 + value1.M42 * value2.M24 + value1.M43 * value2.M34 + value1.M44 * value2.M44;
-
-                return m;
-            }
-        }
+        public static Matrix4x4 operator *(Matrix4x4 value1, Matrix4x4 value2) => Multiply(value1, value2);
 
         /// <summary>
         /// Multiplies a matrix by a scalar value.
@@ -2281,41 +2119,7 @@ namespace System.Numerics
         /// <param name="value1">The source matrix.</param>
         /// <param name="value2">The scaling factor.</param>
         /// <returns>The scaled matrix.</returns>
-        public static unsafe Matrix4x4 operator *(Matrix4x4 value1, float value2)
-        {
-
-            if (Sse.IsSupported)
-            {
-                var all = Sse.SetAllVector128(value2);
-                Sse.Store(&value1.M11, Sse.Multiply(Sse.LoadVector128(&value1.M11), all));
-                Sse.Store(&value1.M21, Sse.Multiply(Sse.LoadVector128(&value1.M21), all));
-                Sse.Store(&value1.M31, Sse.Multiply(Sse.LoadVector128(&value1.M31), all));
-                Sse.Store(&value1.M41, Sse.Multiply(Sse.LoadVector128(&value1.M41), all));
-                return value1;
-            }
-            else
-            {
-                Matrix4x4 m;
-
-                m.M11 = value1.M11 * value2;
-                m.M12 = value1.M12 * value2;
-                m.M13 = value1.M13 * value2;
-                m.M14 = value1.M14 * value2;
-                m.M21 = value1.M21 * value2;
-                m.M22 = value1.M22 * value2;
-                m.M23 = value1.M23 * value2;
-                m.M24 = value1.M24 * value2;
-                m.M31 = value1.M31 * value2;
-                m.M32 = value1.M32 * value2;
-                m.M33 = value1.M33 * value2;
-                m.M34 = value1.M34 * value2;
-                m.M41 = value1.M41 * value2;
-                m.M42 = value1.M42 * value2;
-                m.M43 = value1.M43 * value2;
-                m.M44 = value1.M44 * value2;
-                return m;
-            }
-        }
+        public static Matrix4x4 operator *(Matrix4x4 value1, float value2) => Multiply(value1, value2);
 
         /// <summary>
         /// Returns a boolean indicating whether the given two matrices are equal.
@@ -2323,14 +2127,7 @@ namespace System.Numerics
         /// <param name="value1">The first matrix to compare.</param>
         /// <param name="value2">The second matrix to compare.</param>
         /// <returns>True if the given matrices are equal; False otherwise.</returns>
-        public static bool operator ==(Matrix4x4 value1, Matrix4x4 value2)
-        {
-            return (value1.M11 == value2.M11 && value1.M22 == value2.M22 && value1.M33 == value2.M33 && value1.M44 == value2.M44 && // Check diagonal element first for early out.
-                                                value1.M12 == value2.M12 && value1.M13 == value2.M13 && value1.M14 == value2.M14 &&
-                    value1.M21 == value2.M21 && value1.M23 == value2.M23 && value1.M24 == value2.M24 &&
-                    value1.M31 == value2.M31 && value1.M32 == value2.M32 && value1.M34 == value2.M34 &&
-                    value1.M41 == value2.M41 && value1.M42 == value2.M42 && value1.M43 == value2.M43);
-        }
+        public static bool operator ==(Matrix4x4 value1, Matrix4x4 value2) => value1.Equals(value2);
 
         /// <summary>
         /// Returns a boolean indicating whether the given two matrices are not equal.
@@ -2338,13 +2135,7 @@ namespace System.Numerics
         /// <param name="value1">The first matrix to compare.</param>
         /// <param name="value2">The second matrix to compare.</param>
         /// <returns>True if the given matrices are not equal; False if they are equal.</returns>
-        public static bool operator !=(Matrix4x4 value1, Matrix4x4 value2)
-        {
-            return (value1.M11 != value2.M11 || value1.M12 != value2.M12 || value1.M13 != value2.M13 || value1.M14 != value2.M14 ||
-                    value1.M21 != value2.M21 || value1.M22 != value2.M22 || value1.M23 != value2.M23 || value1.M24 != value2.M24 ||
-                    value1.M31 != value2.M31 || value1.M32 != value2.M32 || value1.M33 != value2.M33 || value1.M34 != value2.M34 ||
-                    value1.M41 != value2.M41 || value1.M42 != value2.M42 || value1.M43 != value2.M43 || value1.M44 != value2.M44);
-        }
+        public static bool operator !=(Matrix4x4 value1, Matrix4x4 value2) => !value1.Equals(value2);
 
         /// <summary>
         /// Returns a boolean indicating whether this matrix instance is equal to the other given matrix.