… src/System.Numerics.Vectors/src/System/Numerics/VectorMath.cs