This PR optimizes some `Matrix4x4` operations with SSE (see https://github.com/dotnet/corefx/issues/31425). Some of the operations could also be optimized with AVX but for some reason on my PC it performs worse than SSE ([VZEROUPPER](https://github.com/dotnet/coreclr/issues/18575)? or maybe CPU decreases AVX frequency due to rapid benchmarking?).

Environment:
```ini
.NET Core: .NET Core SDK=3.0.100-alpha1-20180720-2

Windows 10:
   Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores

macOS 10.13:
   Intel Core i7-4980HQ CPU 2.80GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
```

## Matrix4x4.Add (Matrix4x4, Matrix4x4) and Subtract
`Matrix4x4 result = matrix1 + matrix2;`

Windows (Coffee Lake):

|  Method |      Mean | Scaled |
|-------- |----------:|-------:|
| Add_old | 13.353 ns |   1.00 |
| Add_new |  4.486 ns |   0.34 |

macOS (Haswell):

|  Method |      Mean | Scaled |
|-------- |----------:|-------:|
| Add_old | 15.347 ns |   1.00 |
| Add_new |  7.473 ns |   0.49 |


## Matrix4x4.Lerp (Matrix4x4, Matrix4x4, float)
`Matrix4x4 result = Matrix4x4.Lerp(matrix1, matrix2, amount);`

Windows (Coffee Lake):

|   Method |      Mean | Scaled |
|--------- |----------:|-------:|
| Lerp_old | 15.286 ns |   1.00 |
| Lerp_new |  5.365 ns |   0.35 |

macOS (Haswell):

|   Method |      Mean | Scaled |
|--------- |----------:|-------:|
| Lerp_old | 17.047 ns |   1.00 |
| Lerp_new |  7.657 ns |   0.45 |

## Matrix4x4.Multiply (Matrix4x4, Matrix4x4)
`Matrix4x4 result = matrix1 * matrix2;`

Windows (Coffee Lake):

|       Method |      Mean | Scaled |
|------------- |----------:|-------:|
| Multiply_old | 27.146 ns |   1.00 |
| Multiply_new |  7.461 ns |   0.27 |

macOS (Haswell):

|       Method |     Mean | Scaled |
|------------- |---------:|-------:|
| Multiply_old | 32.05 ns |   1.00 |
| Multiply_new | 11.24 ns |   0.35 |

## Matrix4x4.Multiply (Matrix4x4, float)
`Matrix4x4 result = matrix1 * scalar;`

Windows (Coffee Lake):

|               Method |      Mean | Scaled |
|--------------------- |----------:|-------:|
| MultiplyByScalar_old | 12.927 ns |   1.00 |
| MultiplyByScalar_new |  3.284 ns |   0.25 |

macOS (Haswell):

|               Method |      Mean | Scaled |
|--------------------- |----------:|-------:|
| MultiplyByScalar_old | 14.334 ns |   1.00 |
| MultiplyByScalar_new |  5.086 ns |   0.35 |

## Matrix4x4.Negate (Matrix4x4)
`Matrix4x4 result = -matrix1;`

Windows (Coffee Lake):

|     Method |      Mean | Scaled |
|----------- |----------:|-------:|
| Negate_old | 12.932 ns |   1.00 |
| Negate_new |  3.187 ns |   0.25 |

macOS (Haswell):

|     Method |      Mean | Scaled |
|----------- |----------:|-------:|
| Negate_old | 14.877 ns |   1.00 |
| Negate_new |  5.201 ns |   0.35 |

## Matrix4x4.Equals (Matrix4x4)
`bool result = matrix1 == matrix2;`

Windows (Coffee Lake):

|              Method |     Mean |
|-------------------- |---------:|
| Equals_NotEqual_old | 1.742 ns |
| Equals_NotEqual_new | 1.581 ns |
|    Equals_Equal_old | 7.081 ns |
|    Equals_Equal_new | 2.960 ns |

macOS (Haswell):

|              Method |     Mean |
|-------------------- |---------:|
| Equals_NotEqual_old | 3.172 ns |
| Equals_NotEqual_new | 3.022 ns |
|    Equals_Equal_old | 8.180 ns |
|    Equals_Equal_new | 4.618 ns |

## Matrix4x4.Transpose (Matrix4x4)
`Matrix4x4 result = Matrix4x4.Transpose(matrix1);`

Windows (Coffee Lake):

|        Method |      Mean | Scaled |
|-------------- |----------:|-------:|
| Transpose_old | 12.720 ns |   1.00 |
| Transpose_new |  3.156 ns |   0.25 |

macOS (Haswell):

|        Method |      Mean | Scaled |
|-------------- |----------:|-------:|
| Transpose_old | 14.297 ns |   1.00 |
| Transpose_new |  5.548 ns |   0.39 |

Benchmarks: https://gist.github.com/EgorBo/c80a25517245374c8dcdca2af4536ffe