@tannergooding 2018-08-15
Possibly outside the scope of this PR, but it would be nice if we had some perf tests for `Matrix4x4` (and the other types) in: https://github.com/dotnet/corefx/tree/master/src/System.Numerics.Vectors/tests/Performance