@tannergooding 2018-08-15
CC. @eerhardt, @ViktorHofer 

Also CC. @benaadams, who has looked at some of this before.