@eerhardt 2018-08-16
@EgorBo - the CI is failing because `System.Numerics.Vectors` is built for:

* netcoreapp
* netfx
* netstandard

The Hardware Intrinsics methods you are taking advantage of (i.e. `Vector128<>`) only work/exist on `netcoreapp`. So your code builds fine on that build.

However, for netfx and netstandard, `System.Runtime.Intrinsics` doesn't exist.  So you will need to exclude your new code from the build when building for these configurations.