@tannergooding 2018-08-16
In case people are interested, here are the current PMI disassemblies:
* Software Implementation: [System.Numerics.Vectors.dasm.txt](https://github.com/dotnet/corefx/files/2294452/System.Numerics.Vectors.dasm.txt)
* Hardware Intrinsics Impl: [System.Numerics.Vectors.dasm.txt](https://github.com/dotnet/corefx/files/2294453/System.Numerics.Vectors.dasm.txt)