@CarolEidt 2018-08-16
> Setting COMPlus_FeatureSIMD=0 should cover the Sse-Avx2 HWIntrinsics. I don't think we have a switch to cover all HWIntrinsics, however.

Would it make sense to have a `COMPlus_EnableHWIntrinsics` option that is exposed in Release? It would be set to 1 by default, and could be set to 0 to disable them. Presumably, it would imply `COMPlus_FeatureSIMD` and perhaps the latter should be deprecated?