@tannergooding 2018-08-16
> Would it make sense to have a COMPlus_EnableHWIntrinsics option that is exposed in Release? It would be set to 1 by default, and could be set to 0 to disable them. Presumably, it would imply COMPlus_FeatureSIMD and perhaps the latter should be deprecated?

I think this makes sense, but we probably don't want to deprecate `FeatureSIMD` until all of the types are successfully ported (Matrix4x4 is a case where nothing was Intrinsic, `Vector2/3/4` are the ones that `FeatureSIMD` impacts, IIRC).