@fiigii 2018-08-17
IMO, the current implementation has a very inefficient code pattern that **loads** fields of `Matrix4x4` at the beginning of a function and **stores** the results back at the end. Each operation could generate many memory access operations, so some scenarios that heavily use `Matrix4x4` will possibly become memory bounded.

So, the more efficient implementation should be defining `Matrix4x4` with 4 `Vector128<float>` fields, and other operations directly read the `Vector128<float>` fields then return a `new Matrix4x4(...)`. Because of RyuJIT's struct promotion optimization and inlining, all the loads of local struct's fields and "new" of results will be eliminated (replaced by direct SIMD register operations). According to my experience, that will make some scenarios much faster and more CPU bounded.

cc @tannergooding @CarolEidt