@EgorBo 2018-08-17
@fiigii how would a field accessor look like (e.g. `M11`) - both setter and getter? Will it be interop friendly?