@tannergooding 2018-08-17
@fiigii, I don't believe we can take such a change because the various `System.Numerics.Vector` types were shipped with their fields publicly exposed.

We cannot:
* Remove the existing fields, as it is a major binary breaking change
* Add new fields with explicit layout, as it would
  * Break any existing code using these as interop types (generics can't be marshaled today)
  * Likely regress perf in other areas, due to JIT limitations with explicit layout structs

This also prevents other changes that likely would have been generally good for these types, such as marking the structs as `readonly`, but it is the state of the world and we will need to work with it 😄