@tannergooding 2018-08-17
I would hope that, instead, we could make the JIT smarter for these types of things. Having `__vectorcall` support, for example, would allow the values to be passed around in register and would elide many of the memory concerns.