@tannergooding 2018-08-17
@EgorBo, are there any other changes you were looking to make to this PR?