@EgorBo 2018-08-17
@tannergooding let me switch back to Shuffle-impl for Multiply 🙂