@EgorBo 2018-08-17
I've updated benchmark results in the description with the latest changes, latest runtime and added macOS.
macOS numbers a little bit worse because most of the used intrinsics have better throughput on Coffee Lake (e.g. https://software.intel.com/sites/landingpage/IntrinsicsGuide/#text=_mm_add_ps&expand=127)