@CarolEidt 2018-08-17
@EgorBo - thanks! This is great to see.
@tannergooding @eerhardt - thanks for getting this reviewed and merged. I've commented on a couple of the JIT issues that were raised. Let me know if there are others.